CHG: updating reading queries

+++ src/MSFITSIO.cu
@@ -569,7 +569,7 @@ __host__ void readMS(const char* MS_name,
   data->nstokes = ncorr_col(0);
   casacore::ROArrayColumn<casacore::Int> correlation_col(polarization_tab,
                                                          "CORR_TYPE");
-  casacore::Vector<casacore::Int64> polarizations = correlation_col(0);
+  casacore::Vector<casacore::Int> polarizations = correlation_col(0);
 
   for (int i = 0; i < data->nstokes; i++) {
     data->corr_type.push_back(polarizations[i]);
@@ -580,10 +580,10 @@ __host__ void readMS(const char* MS_name,
   casacore::ROArrayColumn<casacore::Double> chan_freq_col(spectral_window_tab,
                                                           "CHAN_FREQ");
 
-  casacore::ROScalarColumn<casacore::Int64> n_chan_freq(spectral_window_tab,
-                                                        "NUM_CHAN");
+  casacore::ROScalarColumn<casacore::Int> n_chan_freq(spectral_window_tab,
+                                                      "NUM_CHAN");
 
-  casacore::ROScalarColumn<casacore::Int64> spectral_window_ids(
+  casacore::ROScalarColumn<casacore::Int> spectral_window_ids(
       spectral_window_tab, "ID");
 
   for (int i = 0; i < data->n_internal_frequencies; i++) {
